fixes for requested changes
