+++ garden_ai/app/main.py
@@ -9,10 +9,13 @@ import typer
 from garden_ai import GardenClient, GardenConstants
 from garden_ai._version import __version__
 from garden_ai.mcp.config_add import MCPConfigInitalizer as Init
+from garden_ai.app.mcp import mcp_app
 
 logger = logging.getLogger()
 
-app = typer.Typer()
+app = typer.Typer(no_args_is_help=True)
+
+app.add_typer(mcp_app, name="mcp")
 
 
 def show_version(show: bool):
@@ -50,67 +53,8 @@ def logout():
     Path.unlink(Path(GardenConstants.GARDEN_KEY_STORE), missing_ok=True)
 
 
-mcp_app = typer.Typer(help="MCP server commands")
-app.add_typer(mcp_app, name="mcp")
-
-
-@mcp_app.command()
-def serve():
-    """Start the Garden MCP server."""
-    try:
-        from garden_ai.mcp.server import main as mcp_main
-
-        mcp_main()
-    except ImportError:
-        rich.print("[red]Error:[/red] MCP extra not installed.")
-        rich.print("Install with: [cyan]pip install garden-ai[mcp][/cyan]")
-        raise typer.Exit(1)
-
-
-@mcp_app.command()
-def setup(
-    client: Annotated[
-        str | None,
-        typer.Option(help="'claude', 'claude-code', 'gemini', 'cursor', 'windsurf'"),
-    ] = None,
-    path: Annotated[
-        str | None,
-        typer.Option(help="Path to initalize config file for any other mcp client"),
-    ] = None,
-):
-    """Add config file for client"""
-    if client and path:
-        raise ValueError("Cannot specify both a client and a path")
-    elif not client and not path:
-        raise ValueError("Specify either a client or a path")
-
-    if path:
-        config_path = Init.setup_custom(path)
-    elif client:
-        match client.lower():
-            case "claude":
-                config_path = Init.setup_claude()
-            case "claude-code":
-                config_path = Init.setup_claude_code()
-            case "gemini":
-                config_path = Init.setup_gemini()
-            case "cursor":
-                config_path = Init.setup_cursor()
-            case "windsurf":
-                config_path = Init.setup_windsurf()
-            case _:
-                rich.print(
-                    "Not supported for config initialization",
-                    "Try 'claude', 'claude code', 'gemini', 'cursor', or 'windsurf' or specify a path to a config file",
-                )
-                return
-
-    rich.print(f"Garden MCP configuration file set up at {config_path}")
-
-
-@app.callback(invoke_without_command=True)
+@app.callback()
 def main_info(
-    ctx: typer.Context,
     version: Optional[bool] = typer.Option(
         None, "--version", callback=show_version, is_eager=True
     ),
@@ -118,7 +62,4 @@ def main_info(
     """
     🌱 Hello, Garden 🌱
     """
-    if ctx.invoked_subcommand is None:
-        # Manually print the help text if no subcommand is invoked, to get around
-        rich.print(ctx.get_help())
-        raise typer.Exit()
+    pass
