+++ pyproject.toml
@@ -54,7 +54,17 @@ dependencies = [
 mlip = ["ase"]
 mcp = ["mcp>=1.0.0", "uv>=0.8.0"]
 
-test = [
+[dependency-groups]
+develop = [
+    "pre-commit>=3.1.1,<4.0.0",
+    "ruff>=0.6.9,<1.0.0",
+    "mkdocs>=1.4.3,<2.0.0",
+    "mkdocs-callouts>=1.9.0,<2.0.0",
+    "mkdocs-extensions>=0.1.2,<0.2.0",
+    "pymdown-extensions>=10.0.1,<11.0.0",
+    "mkdocs-material>=9.1.13,<10.0.0",
+    "mkdocstrings[python]>=0.25.2,<0.26.0",
+    "jupyter>=1.0.0,<2.0.0",
     "pytest>=7.2.0,<8.0.0",
     "pytest-mock>=3.10.0,<4.0.0",
     "pytest-cov>=4.0.0,<5.0.0",
@@ -71,18 +81,6 @@ test = [
     "pytest-random-order>=1.1.1,<2.0.0",
     "pexpect>=4.9.0,<5.0.0",
     "mixpanel>=4.10.1,<5.0.0",
-]
-
-develop = [
-    "pre-commit>=3.1.1,<4.0.0",
-    "ruff>=0.6.9,<1.0.0",
-    "mkdocs>=1.4.3,<2.0.0",
-    "mkdocs-callouts>=1.9.0,<2.0.0",
-    "mkdocs-extensions>=0.1.2,<0.2.0",
-    "pymdown-extensions>=10.0.1,<11.0.0",
-    "mkdocs-material>=9.1.13,<10.0.0",
-    "mkdocstrings[python]>=0.25.2,<0.26.0",
-    "jupyter>=1.0.0,<2.0.0",
 ]
